use the tooltip API for reporting the coverage
If the sublime 3 version is too old, the coverage file will be opened
(as was the case before this change).
Additionally the user can specify `"pyb_no_popup": true` in his project
configuration to force display of the full report.
The tooltip report contains a link that can be clicked to access the
full text report.